requirements.txt
fuzzywuzzy==0.18.0
pytz==2022.7.1
aiogram==2.25.1
python-dotenv==1.0.0
python-Levenshtein~=0.20.9
ruamel.yaml~=0.17.21